Add @celo/dev-utils
