+++ packages/cli/package.json
@@ -55,6 +55,7 @@
   },
   "devDependencies": {
     "@celo/dev-cli": "^2.0.3",
+    "@celo/dev-utils": "0.0.1",
     "@types/bip32": "^1.0.1",
     "@types/bip39": "^2.4.2",
     "@types/cli-table": "^0.3.0",
