+++ packages/cli/src/commands/account/authorize.test.ts
@@ -1,5 +1,5 @@
+import { testWithGanache } from '@celo/dev-utils/lib/ganache-test'
 import Web3 from 'web3'
-import { testWithGanache } from '../../test-utils/ganache-test'
 import Authorize from './authorize'
 import Register from './register'
