+++ packages/cli/src/commands/account/claims.test.ts
@@ -1,9 +1,9 @@
 import { IdentityMetadataWrapper, newKitFromWeb3 } from '@celo/contractkit'
 import { ClaimTypes } from '@celo/contractkit/lib/identity'
+import { testWithGanache } from '@celo/dev-utils/lib/ganache-test'
 import { readFileSync, writeFileSync } from 'fs'
 import { tmpdir } from 'os'
 import Web3 from 'web3'
-import { testWithGanache } from '../../test-utils/ganache-test'
 import ClaimAccount from './claim-account'
 import ClaimDomain from './claim-domain'
 import ClaimName from './claim-name'
