+++ packages/contractkit/src/identity/claims/account.test.ts
@@ -1,8 +1,8 @@
+import { ACCOUNT_ADDRESSES, ACCOUNT_PRIVATE_KEYS } from '@celo/dev-utils/lib/ganache-setup'
+import { testWithGanache } from '@celo/dev-utils/lib/ganache-test'
 import { privateKeyToAddress, privateKeyToPublicKey } from '@celo/utils/lib/address'
 import { NativeSigner } from '@celo/utils/lib/signatureUtils'
 import { newKitFromWeb3 } from '../../kit'
-import { testWithGanache } from '../../test-utils/ganache-test'
-import { ACCOUNT_ADDRESSES, ACCOUNT_PRIVATE_KEYS } from '../../test-utils/ganache.setup'
 import { IdentityMetadataWrapper } from '../metadata'
 import { createAccountClaim, MetadataURLGetter } from './account'
 import { SignedClaim, verifyClaim } from './claim'
