+++ packages/contractkit/src/test-utils/ganache.teardown.ts
@@ -1,7 +1,2 @@
-export default function tearDown() {
-  console.log('Stopping ganache')
-  return (global as any).stopGanache().catch((err: any) => {
-    console.error('error stopping ganache')
-    console.error(err)
-  })
-}
+import teardown from '@celo/dev-utils/lib/ganache-teardown'
+export default teardown
