+++ packages/contractkit/src/utils/tx-signing.test.ts
@@ -1,10 +1,10 @@
+import { testWithGanache } from '@celo/dev-utils/lib/ganache-test'
 import debugFactory from 'debug'
 import * as util from 'util'
 import Web3 from 'web3'
 import { Provider } from 'web3/providers'
 import { generateAccountAddressFromPrivateKey } from '../providers/celo-private-keys-subprovider'
 import { CeloProvider } from '../providers/celo-provider'
-import { testWithGanache } from '../test-utils/ganache-test'
 import { recoverTransaction } from './signing-utils'
 import { CeloTx } from './tx-signing'
 import { addLocalAccount } from './web3-utils'
