+++ packages/contractkit/src/wrappers/GoldToken.test.ts
@@ -1,5 +1,5 @@
+import { testWithGanache } from '@celo/dev-utils/lib/ganache-test'
 import { newKitFromWeb3 } from '../kit'
-import { testWithGanache } from '../test-utils/ganache-test'
 import { GoldTokenWrapper } from './GoldTokenWrapper'
 
 testWithGanache('GoldToken Wrapper', (web3) => {
