+++ packages/contractkit/src/wrappers/StableToken.test.ts
@@ -1,5 +1,5 @@
+import { testWithGanache } from '@celo/dev-utils/lib/ganache-test'
 import { newKitFromWeb3 } from '../kit'
-import { testWithGanache } from '../test-utils/ganache-test'
 import { StableTokenWrapper } from './StableTokenWrapper'
 
 /*
