+++ packages/dev-utils/package.json
@@ -1,0 +1,28 @@
+{
+  "name": "@celo/dev-utils",
+  "version": "0.0.1",
+  "description": "util package for celo packages that should only be a devDependency",
+  "main": "./lib/index.js",
+  "types": "./lib/index.d.ts",
+  "author": "Celo",
+  "license": "Apache-2.0",
+  "homepage": "https://github.com/celo-org/celo-monorepo/tree/master/packages/dev-utils",
+  "repository": "https://github.com/celo-org/celo-monorepo/tree/master/packages/dev-utils",
+  "keywords": [
+    "celo"
+  ],
+  "scripts": {
+    "build": "tsc -b .",
+    "lint": "tslint -c tslint.json --project ."
+  },
+  "dependencies": {
+    "bignumber.js": "^7.2.0",
+    "web3": "1.0.0-beta.37"
+  },
+  "devDependencies": {
+
+  },
+  "engines": {
+    "node": ">=8.13.0"
+  }
+}
