+++ packages/dev-utils/src/ganache-teardown.ts
@@ -1,0 +1,7 @@
+export default function tearDown() {
+  console.log('Stopping ganache')
+  return (global as any).stopGanache().catch((err: any) => {
+    console.error('error stopping ganache')
+    console.error(err)
+  })
+}
