+++ packages/dev-utils/tsconfig.json
@@ -1,0 +1,9 @@
+{
+  "extends": "../typescript/tsconfig.library.json",
+  "compilerOptions": {
+    "rootDir": "src",
+    "outDir": "lib",
+    "resolveJsonModule": true
+  },
+  "include": ["src", "types/", "src/migration-override.json"]
+}
